other test cleanup
