+++ tests/data/decision_tree/test_projects/geojsons/TEST_01_GRI_04-13-2026.geojson
@@ -1,0 +1,10 @@
+{
+"type": "FeatureCollection",
+"name": "TEST_01_GRI_04-13-2026",
+"crs": { "type": "name", "properties": { "name": "urn:ogc:def:crs:OGC:1.3:CRS84" } },
+"features": [
+{ "type": "Feature", "properties": { "cohort": "terrafund-cohort-2", "project_id": "1826cc5f-0d4d-4427-b5b3-fe244deba919", "poly_id": "22560785-182b-4c06-b5a5-f355eaf4f907", "site_id": "3c2727be-5e1f-4be3-a041-14b6b8bbf97c", "project_name": "TEST_01_GRI", "plantstart": "2022-07-15", "practice": "tree-planting", "target_sys": "riparian-area-or-wetland", "area": 1.8880552734676308, "notes": null }, "geometry": { "type": "Polygon", "coordinates": [ [ [ 30.232814679946795, -2.665400376033629 ], [ 30.232906067029411, -2.66515207206506 ], [ 30.233107118611169, -2.665199541945277 ], [ 30.233172917310657, -2.665068086887885 ], [ 30.233118085061083, -2.665057132299135 ], [ 30.23321312762701, -2.664673721631476 ], [ 30.233322792126145, -2.664494796612369 ], [ 30.233381279859028, -2.664392553732633 ], [ 30.233578675957482, -2.664472887424569 ], [ 30.233684684973323, -2.664330477694365 ], [ 30.234185486186075, -2.664582433359669 ], [ 30.23410872103668, -2.664743100713521 ], [ 30.234141620386421, -2.664750403774561 ], [ 30.234064855237015, -2.664954889466103 ], [ 30.234291495201909, -2.665035223121343 ], [ 30.234393848734445, -2.665079041476539 ], [ 30.234225696502428, -2.665393072976446 ], [ 30.234163553286244, -2.665593907033218 ], [ 30.234218385535822, -2.665648679952104 ], [ 30.234152586836327, -2.665776483420024 ], [ 30.234320739068352, -2.665816650221486 ], [ 30.23424397391895, -2.666053999476089 ], [ 30.232807368980183, -2.665433239790422 ], [ 30.232814679946795, -2.665400376033629 ] ] ] } },
+{ "type": "Feature", "properties": { "cohort": "terrafund-cohort-2", "project_id": "1826cc5f-0d4d-4427-b5b3-fe244deba919", "poly_id": "a09e8e9f-f0f6-438a-95ac-aa5457082d46", "site_id": "3c2727be-5e1f-4be3-a041-14b6b8bbf97c", "project_name": "TEST_01_GRI", "plantstart": "2022-07-15", "practice": "tree-planting", "target_sys": "riparian-area-or-wetland", "area": 0.21031222808946515, "notes": null }, "geometry": { "type": "Polygon", "coordinates": [ [ [ 30.237515353571759, -2.663990986185525 ], [ 30.23781432245924, -2.663999205794307 ], [ 30.237828036628393, -2.663812894648462 ], [ 30.2378472364652, -2.663640282826422 ], [ 30.23788015047116, -2.663530688006222 ], [ 30.237743008779663, -2.663525208264956 ], [ 30.237441297058346, -2.663530688006222 ], [ 30.237433068556847, -2.663536167747464 ], [ 30.237515353571759, -2.663990986185525 ] ] ] } },
+{ "type": "Feature", "properties": { "cohort": "terrafund-cohort-2", "project_id": "1826cc5f-0d4d-4427-b5b3-fe244deba919", "poly_id": "fb738f83-65c2-44cd-9384-ecaa692aff95", "site_id": "3c2727be-5e1f-4be3-a041-14b6b8bbf97c", "project_name": "TEST_01_GRI", "plantstart": "2022-07-15", "practice": "tree-planting", "target_sys": "riparian-area-or-wetland", "area": 0.96218837578182248, "notes": null }, "geometry": { "type": "Polygon", "coordinates": [ [ [ 30.235067591103231, -2.662422678337557 ], [ 30.235241450450076, -2.66234968589219 ], [ 30.235500979620014, -2.662452882106788 ], [ 30.235674838966855, -2.662508255681792 ], [ 30.235891533225253, -2.662508255681792 ], [ 30.236055313769384, -2.66240505947183 ], [ 30.236146022993829, -2.662327033063438 ], [ 30.236030116762592, -2.662110572678809 ], [ 30.235853737715068, -2.662256557593548 ], [ 30.235687437470254, -2.662249006650107 ], [ 30.235158300327676, -2.661851323563456 ], [ 30.234866015048915, -2.661471259227807 ], [ 30.234795463429894, -2.661405817674742 ], [ 30.234571210069475, -2.661521598881646 ], [ 30.234387271919907, -2.661607176288455 ], [ 30.234702234504777, -2.662304380234272 ], [ 30.234906330259776, -2.66222131985709 ], [ 30.235067591103231, -2.662422678337557 ] ] ] } }
+]
+}
